+++ main/src/main/webapp/src/js/index.js
@@ -81,7 +81,7 @@ import {symbolsActionsView} from "./components/views/symbolsActionsView";
 import {symbolsView} from "./components/views/symbolsView";
 import {symbolsHistoryView} from "./components/views/symbolsHistoryView";
 import {symbolsImportView} from "./components/views/symbolsImportView";
-import {symbolsTestView} from './components/views/symbolsTestView';
+import {symbolsTestView} from "./components/views/symbolsTestView";
 import {symbolsTrashView} from "./components/views/symbolsTrashView";
 import {usersSettingsView} from "./components/views/usersSettingsView";
 import {actionForm} from "./components/forms/actions/actionForm";
@@ -126,7 +126,9 @@ import {
     actionFormSendKeys,
     actionFormSubmit,
     actionFormWaitForNode,
-    actionFormWaitForTitle
+    actionFormWaitForTitle,
+    actionFormPressKey,
+    actionFormCheckNodeAttributeValue
 } from "./components/forms/actions/webActionForms";
 import {
     actionFormCall,
@@ -270,6 +272,7 @@ angular
     .component('projectDetailsWidget', projectDetailsWidget)
 
     // web action forms
+    .component('actionFormCheckNodeAttributeValue', actionFormCheckNodeAttributeValue)
     .component('actionFormCheckForNode', actionFormCheckForNode)
     .component('actionFormCheckForText', actionFormCheckForText)
     .component('actionFormCheckPageTitle', actionFormCheckPageTitle)
@@ -284,6 +287,7 @@ angular
     .component('actionFormSubmit', actionFormSubmit)
     .component('actionFormWaitForNode', actionFormWaitForNode)
     .component('actionFormWaitForTitle', actionFormWaitForTitle)
+    .component('actionFormPressKey', actionFormPressKey)
 
     // rest action forms
     .component('actionFormCall', actionFormCall)