+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# ALEX v1.2
+
+## Features
+
+* New actions:
+    * Press special keys like enter, ctrl, etc.
+    * Check the value of an elements attribute
+
 # ALEX v1.1.2
 
 ## Breaking Changes